Move tests into dedicated files
